Reminders Scheduling
Implemented a very basic REST API  for creating and retrieving
reminders.
Created a limited front end page using angular.js to show the stored
reminders
Implemented a cron4j scheduler that sends an email message to a
recipient specified in the reminders.yaml file. The email config is
required for it to work.